-- improve ui of attendance xls upload

+++ app/Http/Controllers/HumanResources/HRDept/AttendanceUploadController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers\HumanResources\HRDept;
 // for controller output
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Redirect;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\DB;
 use Illuminate\View\View;
+
+use Illuminate\Support\Facades\DB;
 use Illuminate\Database\Eloquent\Builder;
 
 // MODELS
@@ -35,6 +36,13 @@ use \Carbon\CarbonInterval;
 use App\Imports\AttendanceImport;
 use Maatwebsite\Excel\Facades\Excel;
 
+// load batch and queue
+use Illuminate\Bus\Batch;
+use Illuminate\Support\Facades\Bus;
+use App\Jobs\AttendancePopulateJob;
+
+use Illuminate\Support\Facades\Storage;
+
 use Session;
 
 class AttendanceUploadController extends Controller
@@ -56,18 +64,42 @@ class AttendanceUploadController extends Controller
 	/**
 	* Show the form for creating a new resource.
 	*/
-	public function create(): View
+	public function create(Request $request): View
 	{
-		return view('humanresources.hrdept.attendance.attendanceupload.create');
+		if (!$request->id) {
+			if (session()->exists('lastBatchIdAttPop')) {
+				$bid = session()->get('lastBatchIdAttPop');
+			} else {
+				$bid = 1;
+			}
+		} else {
+			$bid = $request->id;
+		}
+		$batch = Bus::findBatch($bid);
+		return view('humanresources.hrdept.attendance.attendanceupload.create', ['batch' => $batch]);
 	}
 
 	/**
 	* Store a newly created resource in storage.
 	*/
-	public function store(Request $request): RedirectResponse
+	public function store(Request $request)/*: JsonResponse*/
 	{
+		// $validated = $request->validate(
+		// 	[
+		// 		'softcopy' => 'required|file|max:5120|mimes:xls,csv,xlsx',
+		// 	],
+		// 	[
+		// 		// 'softcopy.required' => 'Please insert year',
+		// 	],
+		// 	[
+		// 		'softcopy' => 'Excel File',
+		// 	]
+		// );
+		if (!$request->file('softcopy')) {
+			return redirect()->route('attendanceupload.create')->with('flash_message', 'No Excel File.');
+			exit;
+		}
 		ini_set('max_execution_time', '0');
-
 		HRTempPunchTime::truncate();
 
 		if ($request->file('softcopy')) {
@@ -98,251 +130,52 @@ class AttendanceUploadController extends Controller
 		}
 
 		// GET THE LATEST ATTENDANCE RECORD DATE IN FACESCAN
-		$query_Recordset3 = DB::select('SELECT DATE(`hr_temp_punch_time`.Att_Time) AS LastDate FROM `hr_temp_punch_time` GROUP BY DATE(`hr_temp_punch_time`.Att_Time) ORDER BY LastDate ASC LIMIT 1');
-		$query_Recordsetnew3 = HRTempPunchTime::groupByRaw('DATE(Att_Time)')->orderBy('Att_Time')->first()->Att_Time;
-		$row_Recordset3 = $query_Recordset3[0]->LastDate;
-		dd($query_Recordsetnew3, $row_Recordset3);
+		// $query_Recordset3 = DB::select('SELECT DATE(`hr_temp_punch_time`.Att_Time) AS LastDate FROM `hr_temp_punch_time` GROUP BY DATE(`hr_temp_punch_time`.Att_Time) ORDER BY LastDate ASC LIMIT 1');
+		$row_Recordset3 = HRTempPunchTime::selectRaw('DATE(`hr_temp_punch_time`.Att_Time) AS LastDate')->groupByRaw('DATE(Att_Time)')->orderBy('Att_Time')->first()->LastDate;
+		// $row_Recordset3 = $query_Recordset3[0]->LastDate;
 
 		// GET THE LATEST ATTENDANCE RECORD DATE IN FACESCAN
-		$query_Recordset5 = DB::select('SELECT DATE(`hr_temp_punch_time`.Att_Time) AS CurrentDate FROM `hr_temp_punch_time` GROUP BY DATE(`hr_temp_punch_time`.Att_Time) ORDER BY CurrentDate DESC LIMIT 1');
-		$row_Recordset5 = $query_Recordset5[0]->CurrentDate;
+		// $query_Recordset5 = DB::select('SELECT DATE(`hr_temp_punch_time`.Att_Time) AS CurrentDate FROM `hr_temp_punch_time` GROUP BY DATE(`hr_temp_punch_time`.Att_Time) ORDER BY CurrentDate DESC LIMIT 1');
+		$row_Recordset5 = HRTempPunchTime::selectRaw('DATE(`hr_temp_punch_time`.Att_Time) AS CurrentDate')->groupByRaw('Att_Time')->orderBy('Att_Time', 'DESC')->first()->CurrentDate;
+		// $row_Recordset5 = $query_Recordset5[0]->CurrentDate;
+		// dd($row_Recordset3, $row_Recordset5);
 
-
-
-
-		// LOOP STAFF FROM ARRAY TO FETCH THE ATTENDANCE
-		foreach ($staffs as $staff) {
-
-			for ($a = strtotime($row_Recordset3); $a <= strtotime($row_Recordset5); $a = strtotime('+1 day', $a)) {
-				$date = date('Y-m-d', $a);
-				$date_name = date('l', $a);
-
-				$totalRows_holiday = HRHolidayCalendar::select('holiday')->where('date_start', '<=', $date)->where('date_end', '>=', $date)->count();
-
-				$totalRows_saturday = HRRestdayCalendar::select('saturday_date')->where('restday_group_id', '=', $staff['restday_group_id'])->where('saturday_date', '=', $date)->count();
-
-				if ($totalRows_holiday > 0) {
-				// HOLIDAY
-					$daytype_id = '3';
-				} elseif ($totalRows_saturday > 0 || $date_name == 'Sunday') {
-				// RESTDAY
-					$daytype_id = '2';
-				} else {
-				// WORKDAY
-					$daytype_id = '1';
-				}
-
-
-				// ----------------------------------------------------------- CONFIGURE IN BREAK RESUME OUT TIME -----------------------------------------------------------//
-				if ($date_name == 'Friday' && $staff['wh_group_id'] == '0') {
-					$row_work_hour = OptWorkingHour::select('time_start_am', 'time_end_am', 'time_start_pm', 'time_end_pm')->where('effective_date_start', '<=', $date)->where('effective_date_end', '>=', $date)->where('group	', '=', $staff['wh_group_id'])->where('category', '=', 3)->first();
-				} elseif ($staff['wh_group_id'] == '0') {
-					$row_work_hour = OptWorkingHour::select('time_start_am', 'time_end_am', 'time_start_pm', 'time_end_pm')->where('effective_date_start', '<=', $date)->where('effective_date_end', '>=', $date)->where('group	', '=', $staff['wh_group_id'])->where('category', '!=', 3)->first();
-				} else {
-					$row_work_hour = OptWorkingHour::select('time_start_am', 'time_end_am', 'time_start_pm', 'time_end_pm')->where('effective_date_start', '<=', $date)->where('effective_date_end', '>=', $date)->where('group	', '=', $staff['wh_group_id'])->where('category', '=', 8)->first();
-				}
-
-
-				// IN
-				$in = "00:00:00";
-				$row_IN = HRTempPunchTime::selectRAW("DATE_FORMAT(Att_Time, '%H:%i:00') AS formatted_time")->where('EmployeeCode', '=', $staff['username'])->whereRaw('DATE(Att_Time) = ?', [$date])->whereRaw('TIME(Att_Time) 	<= ?', [$row_work_hour->time_end_am])->groupBy('formatted_time')->orderBy('Att_Time', 'asc')->first();
-				if ($row_IN != NULL) {
-					$in = $row_IN->formatted_time;
-
-					$in_add = Carbon::parse($in)->addMinutes(1)->format('H:i:s');;
-				}
-
-
-				// OUT
-				$out = "00:00:00";
-				$row_OUT = HRTempPunchTime::selectRAW("DATE_FORMAT(Att_Time, '%H:%i:00') AS formatted_time")->where('EmployeeCode', '=', $staff['username'])->whereRaw('DATE(Att_Time) = ?', [$date])->whereRaw('TIME(	Att_Time) >= ?', [$row_work_hour->time_start_pm])->groupBy('formatted_time')->orderBy('Att_Time', 'desc')->first();
-				if ($row_OUT != NULL) {
-					$out = $row_OUT->formatted_time;
-				}
-
-
-				// BREAK1 (FETCH THE LAST ROW BEFORE BREAK TIME)
-				$break1 = "00:00:00";
-				$row_BREAK1 = HRTempPunchTime::selectRAW("DATE_FORMAT(Att_Time, '%H:%i:00') AS formatted_time")->where('EmployeeCode', '=', $staff['username'])->whereRaw('DATE(Att_Time) = ?', [$date])->whereRaw('TIME(	Att_Time) <= ?', [$row_work_hour->time_end_am])->groupBy('formatted_time')->orderBy('Att_Time', 'desc')->first();
-				if ($row_BREAK1 != NULL) {
-					$break1 = $row_BREAK1->formatted_time;
-				}
-
-
-				// BREAK2 (FETCH THE FIRST ROW BETWEEN BREAK AND RESUME)
-				$break2 = "00:00:00";
-				$break2_difference = "00:00";
-				$row_BREAK2 = HRTempPunchTime::selectRAW("DATE_FORMAT(Att_Time, '%H:%i:00') AS formatted_time")->where('EmployeeCode', '=', $staff['username'])->whereRaw('DATE(Att_Time) = ?', [$date])->whereRaw('TIME(	Att_Time) >= ?', [$row_work_hour->time_end_am])->whereRaw('TIME(Att_Time) <= ?', [$row_work_hour->time_start_pm])->groupBy('formatted_time')->orderBy('Att_Time', 'asc')->first();
-				if ($row_BREAK2 != NULL) {
-					$break2 = $row_BREAK2->formatted_time;
-
-					// CALCULATE INTERVAL TIME BETWEEN PUNCH AND BREAK
-					$break2_date1 = Carbon::parse($date . ' ' . $row_work_hour->time_end_am);
-					$break2_date2 = Carbon::parse($date . ' ' . $row_BREAK2->formatted_time);
-					$break2_interval = $break2_date1->diff($break2_date2);
-					$break2_difference = ($break2_interval->h * 60) + $break2_interval->i;
-				}
-
-
-				// RESUME1 (FETCH THE FIRST ROW AFTER RESUME TIME)
-				$resume1 = "00:00:00";
-				$row_RESUME1 = HRTempPunchTime::selectRAW("DATE_FORMAT(Att_Time, '%H:%i:00') AS formatted_time")->where('EmployeeCode', '=', $staff['username'])->whereRaw('DATE(Att_Time) = ?', [$date])->whereRaw('TIME(	Att_Time) >= ?', [$row_work_hour->time_start_pm])->groupBy('formatted_time')->orderBy('Att_Time', 'asc')->first();
-				if ($row_RESUME1 != NULL) {
-					$resume1 = $row_RESUME1->formatted_time;
-				}
-
-
-				// RESUME2 (FETCH THE LAST ROW BETWEEN BREAK AND RESUME)
-				$resume2 = "00:00:00";
-				$resume2_difference = "00:00";
-				$row_RESUME2 = HRTempPunchTime::selectRAW("DATE_FORMAT(Att_Time, '%H:%i:00') AS formatted_time")->where('EmployeeCode', '=', $staff['username'])->whereRaw('DATE(Att_Time) = ?', [$date])->whereRaw('TIME(	Att_Time) >= ?', [$row_work_hour->time_end_am])->whereRaw('TIME(Att_Time) <= ?', [$row_work_hour->time_start_pm])->groupBy('formatted_time')->orderBy('Att_Time', 'desc')->first();
-				if ($row_RESUME2 != NULL) {
-					$resume2 = $row_RESUME2['formatted_time'];
-
-					// CALCULATE INTERVAL TIME BETWEEN PUNCH AND RESUME
-					$resume2_date1 = Carbon::parse($date . ' ' . $row_work_hour->time_start_pm);
-					$resume2_date2 = Carbon::parse($date . ' ' . $row_RESUME2->formatted_time);
-					$resume2_interval = $resume2_date1->diff($resume2_date2);
-					$resume2_difference = ($resume2_interval->h * 60) + $resume2_interval->i;
-				}
-
-
-				$break = "00:00:00";
-				if ($in != $break1 && $in_add != $break1) {
-					$break = $break1;
-				}
-
-				if ($break2 != $resume2) {
-					$break = $break2;
-				} elseif ($break2_difference < $resume2_difference) {
-					$break = $break2;
-				}
-
-				$resume = "00:00:00";
-				if ($out != $resume1) {
-					$resume = $resume1;
-				}
-
-				if ($resume2 != $break2) {
-					$resume = $resume2;
-				} elseif ($break2_difference >= $resume2_difference) {
-					$resume = $resume2;
-				}
-
-
-				// ----------------------------------------------------------- CONFIGURE WORK HOUR -----------------------------------------------------------//
-				if (($in != '00:00:00' && $break != '00:00:00') || ($in != '00:00:00' && $out != '00:00:00') || ($resume != '00:00:00' && $out != '00:00:00')) {
-
-					// BEGIN TIME
-					if ($in <= $row_work_hour->time_start_am && $in != '00:00:00') {
-						$begin_time = $row_work_hour->time_start_am;
-					} elseif ($in >= $row_work_hour->time_start_am && $in != '00:00:00') {
-						$begin_time = $in;
-					} elseif ($resume <= $row_work_hour->time_start_pm && $resume != '00:00:00') {
-						$begin_time = $row_work_hour->time_start_pm;
-					} elseif ($resume >= $row_work_hour->time_start_pm && $resume != '00:00:00') {
-						$begin_time = $resume;
-					}
-
-					// END TIME
-					if ($out >= $row_work_hour->time_end_pm && $out != '00:00:00') {
-						$end_time = $row_work_hour->time_end_pm;
-					} elseif ($out <= $row_work_hour->time_end_pm && $out != '00:00:00') {
-						$end_time = $out;
-					} elseif ($break >= $row_work_hour->time_end_am && $break != '00:00:00') {
-						$end_time = $row_work_hour->time_end_am;
-					} elseif ($break <= $row_work_hour->time_end_am && $break != '00:00:00') {
-						$end_time = $break;
-					}
-
-					// CALCULATE LUNCH TIME
-					if ($begin_time <= $row_work_hour->time_end_am && $end_time >= $row_work_hour->time_start_pm) {
-						$lunch_break = Carbon::parse($date . ' ' . $row_work_hour->time_end_am);
-						$lunch_resume = Carbon::parse($date . ' ' . $row_work_hour->time_start_pm);
-						$lunch_interval = $lunch_break->diff($lunch_resume);
-						$lunch_difference = ($lunch_interval->h * 60) + $lunch_interval->i;
-					} else {
-						$lunch_difference = '0';
-					}
-
-					// CALCULATE WORK HOUR
-					$begin = Carbon::parse($date . ' ' . $begin_time);
-					$end = Carbon::parse($date . ' ' . $end_time);
-					$total_interval = $begin->diff($end);
-					$total_difference = ($total_interval->h * 60) + $total_interval->i;
-
-					$total_minutes = $total_difference - $lunch_difference;
-					$hours = floor($total_minutes / 60);
-					$minutes = str_pad(($total_minutes % 60), 2, '0', STR_PAD_LEFT);
-					$work_hour = $hours . ":" . $minutes;
-				} else {
-					$work_hour = '00:00:00';
-				}
-
-				// INSERT/UPDATE DATABASE
-				$row_Recordset4 = HRAttendance::select('id', 'in', 'break', 'resume', 'out', 'time_work_hour')->where('attend_date', '=', $date)->where('staff_id', '=', $staff['staff_id'])->first();
-
-				if ($row_Recordset4 != NULL) {
-
-					$new_in = $in;
-					$new_break = $break;
-					$new_resume = $resume;
-					$new_out = $out;
-					$new_work_hour = $work_hour;
-
-
-					if ($row_Recordset4->in != '00:00:00' && $in == '00:00:00') {
-						$new_in = $row_Recordset4->in;
-					} elseif ($row_Recordset4->in != '00:00:00' && $in != '00:00:00' && $row_Recordset4->in <= $in) {
-						$new_in = $row_Recordset4->in;
-					}
-
-					if ($row_Recordset4->break != '00:00:00' && $break == '00:00:00') {
-						$new_break = $row_Recordset4->break;
-					} elseif ($row_Recordset4->break != '00:00:00' && $break != '00:00:00' && $row_Recordset4->break <= $break) {
-						$new_break = $row_Recordset4->break;
-					}
-
-					if ($row_Recordset4->resume != '00:00:00' && $resume == '00:00:00') {
-						$new_resume = $row_Recordset4->resume;
-					} if ($row_Recordset4->resume != '00:00:00' && $resume != '00:00:00' && $row_Recordset4->resume >= $resume) {
-						$new_resume = $row_Recordset4->resume;
-					}
-
-					if ($row_Recordset4->out != '00:00:00' && $out == '00:00:00') {
-						$new_out = $row_Recordset4->out;
-					} elseif ($row_Recordset4->out != '00:00:00' && $out != '00:00:00' && $row_Recordset4->out >= $out) {
-						$new_out = $row_Recordset4->out;
-					}
-
-					if ($row_Recordset4->time_work_hour != '' && $row_Recordset4->time_work_hour != NULL && $row_Recordset4->time_work_hour >= $work_hour) {
-						$new_work_hour = $row_Recordset4->time_work_hour;
-					}
-
-					HRAttendance::find($row_Recordset4->id)->update([
-						'in' => $new_in,
-						'break' => $new_break,
-						'resume' => $new_resume,
-						'out' => $new_out,
-						'time_work_hour' => $new_work_hour,
-					]);
-				} else {
-					HRAttendance::create([
-						'staff_id' => $staff['staff_id'],
-						'daytype_id' => $daytype_id,
-						'attend_date' => $date,
-						'in' => $in,
-						'break' => $break,
-						'resume' => $resume,
-						'out' => $out,
-						'time_work_hour' => $work_hour,
-					]);
-				}
+		$data = array_chunk($staffs, 2);
+		foreach ($data as $index => $values) {
+			// $datacsv[$index] = $values;
+			foreach ($values as $dataval) {
+				// combine header with data and pickup column that we need
+				$datacsv[$index][] = $dataval;
 			}
-		}
+			// dd($datacsv[$index]);
 
-		Session::flash('flash_message', 'Successfully upload excel.');
-		return redirect()->route('attendanceupload.create');
+			// call queues by chunk
+			// AttendancePopulateJob::dispatch($datacsv[$index]);
+
+			// we need a progress so we use batch n comment out the queue above
+			$dat[] = new AttendancePopulateJob($datacsv[$index], $row_Recordset3, $row_Recordset5);
+		}
+		// dd(new AttendancePopulateJob($datacsv[$index], $row_Recordset3, $row_Recordset5), $dat);
+
+		$batches = Bus::batch($dat)
+					->name('Attendance Populate Job on -> '.now()->format('j F Y g:i:s a'))
+					->progress(function (Batch $batch) {
+						// A single job has completed successfully...
+					})
+					->then(function (Batch $batch) {
+						// All jobs completed successfully...
+					})
+					->catch(function (Batch $batch, Throwable $e) {
+						// First batch job failure detected...
+					})
+					->finally(function (Batch $batch) {
+						// The batch has finished executing...
+					})
+					->dispatch();
+
+		Session(['lastBatchIdAttPop' => $batches->id]);
+		// return redirect()->route('attendanceupload.create', ['id' => $batches->id])->with('flash_message', 'Successfully upload excel.');
+		return response()->json(route('attendanceupload.create', ['id' => $batches->id]));
 	}
 
 	/**
